requirements.txt
flask
numpy
pandas
pinecone_client
pymongo
scikit_learn
flask_cors
